Disk drive; Measurement works

+++ Inc/ms5534.h
@@ -2,7 +2,7 @@
 #define __MS5354_H_
 void MS55Init(void);
 void MS55Reset(void);
-void MS5534Measure(double * temp, double * pressure);
+void MS5534Measure(float * temp, float * pressure);
 
 
 #endif
